DEV-574 #time 1m minor fix for S3R to allow for larger BT TX buffers

+++ Comms/shimmer_bt_uart.c
@@ -2488,10 +2488,10 @@ void ShimBt_sendNextChar(void)
     BtTransmit(&buf, 1);
 #else
     HAL_StatusTypeDefShimmer ret_val;
-    uint8_t numBytes;
+    uint16_t numBytes;
 
-    uint8_t rdIdx = (gBtTxFifo.rdIdx & BT_TX_BUF_MASK);
-    uint8_t wrIdx = (gBtTxFifo.wrIdx & BT_TX_BUF_MASK);
+    uint16_t rdIdx = (gBtTxFifo.rdIdx & BT_TX_BUF_MASK);
+    uint16_t wrIdx = (gBtTxFifo.wrIdx & BT_TX_BUF_MASK);
 
     if (rdIdx < wrIdx)
     {
